+++ nyaacli/torrenting.py
@@ -3,15 +3,14 @@ import sys
 import time
 import logging
 
+from rich.progress import Progress, BarColumn, DownloadColumn, TransferSpeedColumn, TextColumn, TimeRemainingColumn
+from rich import print
 import libtorrent
-
-from nyaacli.colors import red, green
-from nyaacli.utils import clear_screen
 
 logger = logging.getLogger('nyaa')
 
 
-def download_torrent(filename: str, result_filename: str = None, show_progress: bool = True, base_path: str = 'Anime'):
+def download_torrent(filename: str, result_filename: str = None, show_progress: bool = True, base_path: str = 'Anime') -> str:
     session = libtorrent.session({'listen_interfaces': '0.0.0.0:6881'})
     logger.debug('Started libtorrent session')
 
@@ -28,34 +27,53 @@ def download_torrent(filename: str, result_filename: str = None, show_progress: 
 
     status: libtorrent.session_status = handle.status()
 
-    print(f"\n{green('◉ Started downloading torrent:')} {status.name}", end='\n\n')
+    progress_bar = Progress(
+        '[progress.description]{task.description}',
+        BarColumn(bar_width=None),
+        '[progress.percentage]{task.percentage:>3.1f}%',
+        '•',
+        DownloadColumn(),
+        '•',
+        TransferSpeedColumn(),
+        '•',
+        TimeRemainingColumn(),
+        '•',
+        TextColumn('[green]Peers: {task.fields[peers]}[/green]')
+    )
 
     if show_progress:
-        while not status.is_seeding:
-            status = handle.status()
+        with progress_bar:
+            download_task = progress_bar.add_task('downloading', filename=status.name, total=status.total_wanted, peers=0, start=False)
 
-            progress = green(f"{status.progress * 100:.2f}%")
-            download_rate = green(f"{status.download_rate / 1000:.1f} kB/s")
+            while not status.total_done:
+                # Checking files
+                status = handle.status()
+                description = '[bold yellow]Checking files[/bold yellow]'
+                progress_bar.update(download_task, completed=status.total_done, peers=status.num_peers, description=description)
 
-            print(
-                f'◉ {green(str(status.state).title())} - {progress} '
-                f'(Download: {download_rate} - Upload: {status.upload_rate / 1000:.1f} kB/s - '
-                f'Peers: {status.num_peers}) ',
-                end='\r'
+            # Started download
+            progress_bar.start_task(download_task)
+            description = f'[bold blue]Downloading[/bold blue] [bold yellow]{result_filename}[/bold yellow]'
+
+            while not status.is_seeding:
+                status = handle.status()
+
+                progress_bar.update(download_task, completed=status.total_done, peers=status.num_peers, description=description)
+
+                alerts = session.pop_alerts()
+
+                alert: libtorrent.alert
+                for alert in alerts:
+                    if alert.category() & libtorrent.alert.category_t.error_notification:
+                        logger.debug(f"[Alert] {alert}")
+
+                time.sleep(1)
+
+            progress_bar.update(
+                download_task,
+                description=f'[bold blue]Finished Downloading[/bold blue] [bold green]{result_filename}[/bold green]',
+                completed=status.total_wanted
             )
-
-            alerts = session.pop_alerts()
-
-            alert: libtorrent.alert
-            for alert in alerts:
-                if alert.category() & libtorrent.alert.category_t.error_notification:
-                    logger.debug(f"{red('[Alert]')} {alert}")
-
-            sys.stdout.flush()
-
-            time.sleep(1)
-
-        print(green(f'\n\nFinished downloading {handle.name()}'), end='\n\n')
 
     if result_filename:
         old_name = f'{base_path}/{status.name}'
@@ -63,10 +81,11 @@ def download_torrent(filename: str, result_filename: str = None, show_progress: 
 
         os.rename(old_name, new_name)
 
-        clear_screen()
+        logger.debug(f"Finished torrent download, renamed '{old_name}' to '{new_name}'")
 
-        logger.debug('Finished torrent download')
-        print(f'Finished download at: \'{green(new_name)}\' ')
+        return new_name
+
+    return ''
 
 
 if __name__ == '__main__':